requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
motor==3.5.1
httpx==0.27.2
feedparser==6.0.11
beautifulsoup4==4.12.3
jinja2==3.1.4
python-dateutil==2.9.0.post0
apscheduler==3.10.4
